error handling and comments of the backend

+++ src/backend/controllers/userContoller.js
@@ -1,17 +1,11 @@
 import User from "../models/userModel";
 import { uploads } from "../utils/cloudnary";
-import AppError from "../utils/errClass";
 import fs from "fs";
 import bcrypt from "bcryptjs";
-import formidable from "formidable";
 
 export const registerUser = async (req, res, next) => {
   const { name, email, password, role } = req.body;
   const newUser = await User.create({ name, email, password, role });
-
-  if (!newUser) {
-    return next(new AppError());
-  }
 
   // return NextResponse.json({ newProduct });
   res.status(201).json({ newUser });
@@ -25,51 +19,52 @@ export const getAllUsers = async (req, res, next) => {
 };
 
 export const updateUser = async (req, res, next) => {
-  console.log("files------>", req.files);
-  const file = req.files[0];
-
-  const { path } = file;
-  console.log("path------>", path);
-
-  // const formData = await req.formData();
-
-  // const file = formData.get("file");
-
-  // console.log("files------>", req.files);
-
   const newUserData = {
     name: req.body.name,
     email: req.body.email,
   };
+  //CHECK IF THE USER EXISTS AND SEND A CORRESPONDING MESSAGE IF NOT FOUND
+  let updatedUser = await User.findById(req.user?.id);
 
+  if (!updatedUser) {
+    return res.status(404).json({ message: "User not found" });
+  }
+
+  //UPLOAD THE IMAGE TO CLOUDINARY ANAD DELETE THE IMAGE FROM THE FILE SYSTEM SUING THE FS MODULE
   if (req.files.length > 0) {
     const uploader = async (path) =>
       await uploads(path, "egadgetsApp/userPhotos");
 
     const file = req.files[0];
 
-    console.log(file);
-
     const { path } = file;
 
     const avatarResponse = await uploader(path);
-    console.log(avatarResponse);
+
     fs.unlinkSync(path);
     newUserData.avatar = avatarResponse;
   }
 
-  // console.log(newUserData);
-  const updatedUser = await User.findByIdAndUpdate(req.user._id, newUserData);
+  //UPDATE THE USER AND SEND RESPONSE TO THE FRONTEND
+  updatedUser = await User.findByIdAndUpdate(req.user._id, newUserData);
 
   res.status(200).json({
-    status: "SUCESS",
-    // updatedUser,
+    success: true,
+    updatedUser,
   });
 };
 
 export const updatePassword = async (req, res, next) => {
+  //CHECK IF THE USER EXISTS AND SEND A CORRESPONDING MESSAGE IF NOT FOUND
   const user = await User.findById(req.user._id).select("+password");
 
+  if (!user) {
+    return res
+      .status(400)
+      .json({ message: "please login again to perform this action" });
+  }
+
+  //COMPARE THE INPUTTED PASSWORD AND THE DATABASE PASSWORD USING THE BCRYPT PACKAGE, IF THEY DONT MATCH SEND A CORRESPONDING MESSGAE
   const isPasswordMatch = await bcrypt.compare(
     req.body.currentPassword,
     user.password
@@ -79,6 +74,7 @@ export const updatePassword = async (req, res, next) => {
     return res.status(400).json({ message: "password does not match" });
   }
 
+  //UPDATE THE USER PASSWORD AND SAVE THE DOCUMENT SO THAT THE PRESAVE MIDDLEWARE CAN WORK SO AS TO ENCRPT THE PASSWORD BEFORE SAVING TO DATABASE
   user.password = req.body.newPassword;
 
   await user.save();
